+++ ChangeLog.md
@@ -1,6 +1,7 @@
 # COFantasy: Historique des changements (depuis la version 1.0)
 ## 3.02
 ### Capacité
+* Ajout des réaction allergiques pour les chiens gobelins
 * Possibilité optionelle de lancer plus de soins légers en une journée que le rang dans la voie, en échange d'un coût en mana.
 
 ### Corrections de bugs